mqtt: Add connection to Mosquitto broker
This commit implements the connection to a mosquitto broker to exchange messages for
the servo motor control and logging purposes. This initial commit only includes the
connection to a localhost broker and hard-coded provided topics.
Mosquitto lib required linkage to the target is specified, though dependencies to the
mosquitto package shall be defined and manage by yocto (DEPENDS to be defined in the recipe).

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>