mqtt: Add rum-time configuration for change prone data
This commit implements a place-holder class for mosquitto change prone data, such as host.
The configuration is parsed at application boot and enables to update change prone data without
having to rebuild the application, only the json config file needs to be updated. The target
file path (careful - the path given must be the file location in the generated rootfs)
should be defined in yocto and passed to CMake through the EXTRA_OECMAKE feature.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>