servo: Implement servo motor interface
This commit implements the servo motor interface which interacts with the dedicated
kernel module. The application uses IOCTL syscalls to initialize the pwm channel used
by the servo motor.

Moreover, the pwm channel run-time configuration is also implemented in this commit.
The configuration allows the specification of the pwm channel, its period and the
kernel module character device.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>